Give node split priority over other leaf work (#369)
* Give split priority over other leaf work.


* Use leaf iteration macros to walk leaf nodes.


* Remove level-base caching.


* Unconditionally cache hblocks except at cn_open() time.
Avoid kvsets with large key counts during length remediation and idle node compaction.
Allow up to three split threads to run concurrently (on different kvs').

+++ lib/include/hse_ikvdb/kvset_view.h
@@ -12,9 +12,9 @@ struct cn;
 struct kvset;
 
 struct kvset_metrics {
-    uint32_t num_keys;
-    uint32_t num_tombstones;
-    uint32_t nptombs;
+    uint64_t num_keys;
+    uint64_t num_tombstones;
+    uint64_t nptombs;
     uint32_t num_hblocks; /* this should always be one */
     uint32_t num_kblocks;
     uint32_t num_vblocks;
